11  - Capstone project/requirements.txt
streamlit==1.36.0
openai==1.37.0
requests==2.32.3
python-dotenv==1.0.1
httpx<0.28
